Add method to check if Piwik URL is correct
Happy to change the name, used ping similar to eg https://redis.io/commands/ping . We could maybe also rename it to eg `testUrl()` or `verifyUrl()` as `ping()` is confusing with `doPing()`. There are many ways that can be used for it. 

Ideally we would check for response code instead of checking for word "piwik" but the method `sendRequest` does not return the response code.

Ideally, it would also test whether the configured idSite actually exists but so far I have not figured out how to do it without changing Piwik. eg "https://demo.piwik.org/piwik.php?idsite=9999&rec=1" would return an error and we could be sure the idSite does not exist, but it is recording an actual request. "https://demo.piwik.org/piwik.php?idsite=9999&rec=0" always returns a valid image with HTTP 200 so we cannot test whether the site exists. Any ideas?